BUG: subj_idx in kabuki_data_to_hddm_data was float instead of int

+++ hddm/generate.py
@@ -159,7 +159,7 @@ def kabuki_data_to_hddm_data(kabuki_data):
     "transform data generate by kabuki.generate.gen_rand_data to hddm data"
 
     kd = kabuki_data
-    dtype = ([('response', np.int), ('rt', np.float), ('subj_idx', np.float), ('condition', 'S20')])
+    dtype = ([('response', np.int), ('rt', np.float), ('subj_idx', np.int32), ('condition', 'S20')])
     data = np.empty(kd.shape, dtype = dtype)
 
     rts = kd['data']
